requirements.txt
anyio==4.11.0
click==8.3.0
fastapi==0.110.3
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==1.10.24
PyMySQL==1.1.2
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.37.2
swagger_ui_bundle==1.1.0
typing_extensions==4.15.0
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
pydantic[email]
